add program 102 (14)

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,15 +1,17 @@
 #include "main.h"
 #include <stdio.h>
 
-/*
- * function main prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line
+/**
+ * main - print the first 50 fibonacci numbers
+ *
+ * Return: Nothing.
  */
 
 int main(void)
 {
-	long long int a = 1;
-	long long int b = 2;
-	long long int c;
+	double a = 1;
+	double b = 2;
+	double c;
 	int i;
 
 	printf("%i, %i", a, b);
